+++ packages/adapter-session/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @swr-login/adapter-session
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @swr-login/core@0.5.0
+
 ## 0.2.2
 
 ### Patch Changes